+++ changelog.md
@@ -1,3 +1,12 @@
+## 23.4.0
+- added color to buttonGroup
+- added buttonContext
+- buttonGroup use buttonContext as provider
+- added color to tabs
+- tabs use buttonContext as provider
+- fixed drawer temporary position
+- change mobile detection from 1024 to 640px
+
 ## 23.3.2
 - style fixes